GMS-918: Update preset contract SDK with batch minting changes.
This commit is based off of the [preset contract changes to support
batch
minting](https://github.com/immutable/zkevm-contracts/pull/47/files).
As a result, the following methods were removed because the contract is
no longer enumerable:

* `tokenByIndex`
* `tokenOfOwnerByIndex`

A number of methods were added or modified to support token ID
specification, and/or batch minting/burning/transfer functionality.