Ensure balance carries over when calculating requirements
Until now, a given token (ERC20 or native) would always have a single
requirement line. Since fee requirements are not being aggregated with
non-fees now, the logic for calculating if the requirements will be met
needs to change.

Instead of calculating against the user balance every time, we'll
similate a balance change before moving on to the next calculation. This
way, any computed changes will carry over, thus failing if the balance
isn't enough after it's been deducted by a previous requirement.

As part of this change, the getTokenBalanceRequirement method has been
made sync, so using its result inside a loop doesn't require an await.
The async part of this method has been moved to a different method.